refactor: move indexOn search logic into separate module
